Fix Is.String* to Does.* for exception expectancy
Is.StringContaining of Is.StringMatching is deprecated in v3, so
migration needs to convert to a proper API (e.g. Does.Contain).

VerifyCSharpFix(source, expected, allowNewCompilerDiagnostics: true) was
used to suppress compilation errors in unit tests when code has been
migrated to v3 API that breaks v2. This catches all other errors and
disregard them which is dangerous, so a proper way of exclusions needs
to be provided. See #10.

Added previously omitted case of MessageMatch.StartsWith.

Closes #1.